Moved concern of tracking aggregateId and version outward
Within an aggregate instance it should not be necessary to pass around the id and version. It is what it is after all.

But.. when dealing with storing events and/or sending commands it is important to know the aggregate id and version.

For convenvenience the aggregateId and version are combined into a 'VersionedId'

Still to do is locking and/or detecting concurrent modifications.