Storing aggregate types with events
This is useful when storing events for different aggregate types in a single
event store (as is done here). Specifically because of the way aggregates are
reloaded. It's good to have *both* the aggregate type and UUID when doing this,
to prevent that events of the wrong type are being applied.

This approach will make it easier when CAs can have either a TA or another CA
as parent. The child CA just knows its parent by UUID. This way we can try to
load this UUID as a TA if there is NONE (rather than a bunch of exceptions for
unknown events), we can try loading a CA with that UUID.